requirements.txt
selenium
webdriver-manager
pyyaml
python-dotenv